Move loadDocument() to abstract class again
