+++ Classes/Controller/AbstractController.php
@@ -14,6 +14,7 @@ namespace Kitodo\Dlf\Controller;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 
+use Kitodo\Dlf\Common\Document;
 
 /**
  *
@@ -21,4 +22,61 @@ use Psr\Log\LoggerAwareTrait;
 abstract class AbstractController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController implements LoggerAwareInterface
 {
     use LoggerAwareTrait;
+
+    /**
+     * Loads the current document into $this->doc
+     *
+     * @access protected
+     *
+     * @return void
+     */
+    protected function loadDocument($requestData)
+    {
+        // Check for required variable.
+        if (
+            !empty($requestData['id'])
+            && !empty($this->settings['pages'])
+        ) {
+            // Should we exclude documents from other pages than $this->settings['pages']?
+            $pid = (!empty($this->settings['excludeOther']) ? intval($this->settings['pages']) : 0);
+            // Get instance of \Kitodo\Dlf\Common\Document.
+            $this->doc = Document::getInstance($requestData['id'], $pid);
+            if (!$this->doc->ready) {
+                // Destroy the incomplete object.
+                $this->doc = null;
+                $this->logger->error('Failed to load document with UID ' . $requestData['id']);
+            } else {
+                // Set configuration PID.
+                $this->doc->cPid = $this->settings['pages'];
+            }
+        } elseif (!empty($requestData['recordId'])) {
+            $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
+                ->getQueryBuilderForTable('tx_dlf_documents');
+
+            // Get UID of document with given record identifier.
+            $result = $queryBuilder
+                ->select('tx_dlf_documents.uid AS uid')
+                ->from('tx_dlf_documents')
+                ->where(
+                    $queryBuilder->expr()->eq('tx_dlf_documents.record_id', $queryBuilder->expr()->literal($requestData['recordId'])),
+                    Helper::whereExpression('tx_dlf_documents')
+                )
+                ->setMaxResults(1)
+                ->execute();
+
+            if ($resArray = $result->fetch()) {
+                $requestData['id'] = $resArray['uid'];
+                // Set superglobal $_GET array and unset variables to avoid infinite looping.
+                $_GET[$this->prefixId]['id'] = $requestData['id'];
+                unset($requestData['recordId'], $_GET[$this->prefixId]['recordId']);
+                // Try to load document.
+                $this->loadDocument();
+            } else {
+                $this->logger->error('Failed to load document with record ID "' . $requestData['recordId'] . '"');
+            }
+        } else {
+            $this->logger->error('Invalid UID ' . $requestData['id'] . ' or PID ' . $this->settings['pages'] . ' for document loading');
+        }
+    }
+
 }
