+++ Classes/Controller/TableOfContentsController.php
@@ -12,13 +12,12 @@
 namespace Kitodo\Dlf\Controller;
 
 use Kitodo\Dlf\Common\Document;
+use Kitodo\Dlf\Common\Helper;
+
 use TYPO3\CMS\Core\Configuration\ExtensionConfiguration;
 use TYPO3\CMS\Core\Utility\MathUtility;
-use \TYPO3\CMS\Extbase\Configuration\ConfigurationManager;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use Kitodo\Dlf\Common\Helper;
 use TYPO3\CMS\Core\Database\ConnectionPool;
-use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
 
 /**
  * Controller for plugin 'Table Of Contents' for the 'dlf' extension
@@ -28,7 +27,7 @@ use TYPO3\CMS\Extbase\Mvc\Controller\ActionController;
  * @subpackage dlf
  * @access public
  */
-class TableOfContentsController extends ActionController
+class TableOfContentsController extends AbstractController
 {
     protected $prefixId = 'tx_dlf';
 
@@ -46,16 +45,6 @@ class TableOfContentsController extends ActionController
     protected $extensionConfiguration;
 
     /**
-     * @var ConfigurationManager
-     */
-    protected $configurationManager;
-
-    /**
-     * @var \TYPO3\CMS\Core\Log\LogManager
-     */
-    protected $logger;
-
-    /**
      * @var array
      */
     protected $pluginConf;
@@ -65,8 +54,6 @@ class TableOfContentsController extends ActionController
      */
     public function __construct()
     {
-        $this->configurationManager = GeneralUtility::makeInstance(ConfigurationManager::class);
-        $this->logger = GeneralUtility::makeInstance('TYPO3\CMS\Core\Log\LogManager')->getLogger(__CLASS__);
         $this->extensionConfiguration = GeneralUtility::makeInstance(ExtensionConfiguration::class)->get('dlf');
 
         // Read plugin TS configuration.
@@ -309,59 +296,4 @@ class TableOfContentsController extends ActionController
     }
 
     // TODO: Needs to be placed in an abstract class
-    /**
-     * Loads the current document into $this->doc
-     *
-     * @access protected
-     *
-     * @return void
-     */
-    protected function loadDocument($requestData)
-    {
-        // Check for required variable.
-        if (
-            !empty($requestData['id'])
-            && !empty($this->settings['pages'])
-        ) {
-            // Should we exclude documents from other pages than $this->settings['pages']?
-            $pid = (!empty($this->settings['excludeOther']) ? intval($this->settings['pages']) : 0);
-            // Get instance of \Kitodo\Dlf\Common\Document.
-            $this->doc = Document::getInstance($requestData['id'], $pid);
-            if (!$this->doc->ready) {
-                // Destroy the incomplete object.
-                $this->doc = null;
-                $this->logger->error('Failed to load document with UID ' . $requestData['id']);
-            } else {
-                // Set configuration PID.
-                $this->doc->cPid = $this->settings['pages'];
-            }
-        } elseif (!empty($requestData['recordId'])) {
-            $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)
-                ->getQueryBuilderForTable('tx_dlf_documents');
-
-            // Get UID of document with given record identifier.
-            $result = $queryBuilder
-                ->select('tx_dlf_documents.uid AS uid')
-                ->from('tx_dlf_documents')
-                ->where(
-                    $queryBuilder->expr()->eq('tx_dlf_documents.record_id', $queryBuilder->expr()->literal($requestData['recordId'])),
-                    Helper::whereExpression('tx_dlf_documents')
-                )
-                ->setMaxResults(1)
-                ->execute();
-
-            if ($resArray = $result->fetch()) {
-                $requestData['id'] = $resArray['uid'];
-                // Set superglobal $_GET array and unset variables to avoid infinite looping.
-                $_GET[$this->prefixId]['id'] = $requestData['id'];
-                unset($requestData['recordId'], $_GET[$this->prefixId]['recordId']);
-                // Try to load document.
-                $this->loadDocument();
-            } else {
-                $this->logger->error('Failed to load document with record ID "' . $requestData['recordId'] . '"');
-            }
-        } else {
-            $this->logger->error('Invalid UID ' . $requestData['id'] . ' or PID ' . $this->settings['pages'] . ' for document loading');
-        }
-    }
 }
