+++ ext_localconf.php
@@ -56,12 +56,6 @@ plugin.tx_dlf_collection {
 }
 tt_content.list.20.dlf_collection < plugin.tx_dlf_collection
 
-plugin.tx_dlf_feeds = USER_INT
-plugin.tx_dlf_feeds {
-    userFunc = Kitodo\Dlf\Plugin\Feeds->main
-}
-tt_content.list.20.dlf_feeds < plugin.tx_dlf_feeds
-
 plugin.tx_dlf_listview = USER_INT
 plugin.tx_dlf_listview {
     userFunc = Kitodo\Dlf\Plugin\ListView->main
@@ -109,12 +103,6 @@ plugin.tx_dlf_statistics {
     userFunc = Kitodo\Dlf\Plugin\Statistics->main
 }
 tt_content.list.20.dlf_statistics < plugin.tx_dlf_statistics
-
-plugin.tx_dlf_tableofcontents = USER
-plugin.tx_dlf_tableofcontents {
-    userFunc = Kitodo\Dlf\Plugin\TableOfContents->main
-}
-tt_content.list.20.dlf_tableofcontents < plugin.tx_dlf_tableofcontents
 
 plugin.tx_dlf_toolbox = USER
 plugin.tx_dlf_toolbox {
